% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_funs.R
\name{clean_ctl}
\alias{clean_ctl}
\title{FUNCTION_TITLE}
\usage{
clean_ctl(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[set_names][purrr::set_names]
}
