% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_template.R
\name{ctl_block}
\alias{ctl_block}
\title{FUNCTION_TITLE}
\usage{
ctl_block(BLOCK)
}
\arguments{
\item{BLOCK}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
}
