% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_template.R
\name{ctl_blocks}
\alias{ctl_blocks}
\title{FUNCTION_TITLE}
\usage{
ctl_blocks(BLOCKS)
}
\arguments{
\item{BLOCKS}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}},\code{\link[purrr]{set_names}}
}
