% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_compare.R
\name{ctl_compare}
\alias{ctl_compare}
\title{Visually compare CTLs}
\usage{
ctl_compare(ctl1, ctl2)
}
\arguments{
\item{ctl1}{character, lines of CTL (left hand side)}

\item{ctl2}{character, lines of CTL (right hand side)}
}
\value{
htmlwidget
}
\description{
Visually compare two CTL files
}
\seealso{
\code{\link[diffr]{diffr}}
}
