% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_template.R
\name{ctl_create}
\alias{ctl_create}
\title{FUNCTION_TITLE}
\usage{
ctl_create(template, ...)
}
\arguments{
\item{template}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[whisker]{whisker.render}}
 \code{\link[purrr]{flatten}}
}
