% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_funs.R
\name{ctl_to_mat}
\alias{ctl_to_mat}
\title{Parse CTL blocks into matrices}
\usage{
ctl_to_mat(x, type = c("numeric", "comment"))
}
\arguments{
\item{x}{clean_ctl object}

\item{type}{character, extract numeric of comment from x, Default: c("numeric", "comment")}
}
\value{
matrix or list containing matrices
}
\description{
Convert CTL blocks into matrices or matrices fo comments
}
\details{
expected format for comment in CTL is "VALUE [FIXED,SAME] ;  [COVTYPE] LABEL"
}
\examples{
\dontrun{
project <- system.file('streams',package = 'streamline')

run <- '101'

xml_raw   <- read_nmlist(run, project)

x   <- xml_raw\%>\%
  xml2::xml_find_first('.//nm:control_stream')\%>\%
  xml2::xml_text()\%>\%
  clean_ctl()

ctl_to_mat(x$OMEGA)
ctl_to_mat(x$OMEGA,type='comment')
}
}
