% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_template.R
\name{ctl_wrap}
\alias{ctl_wrap}
\title{FUNCTION_TITLE}
\usage{
ctl_wrap(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
