% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_funs.R
\name{parse_theta}
\alias{parse_theta}
\title{Parse $THETA into tabular form}
\usage{
parse_theta(x)
}
\arguments{
\item{x}{list of character vectors making up the $THETA section of a control stream}
}
\value{
tibble containing: variable number (i.e., THETA_i, i=1,...,N); FIXED, indicating presence of FIX or FIXED in the initial condition; LB, INITITAL, and UB describing the initial condition itself; and UNIT and LABEL which are parsed from the comment.
}
\description{
Parse $THETA from the control stream.
}
\details{
Parse $THETA into a form that shows initial values and fixed or not fixed. They are created from the comment and follows the form \code{<init> ; [UNIT] LABEL}.
}
\seealso{
[stri_count][stringi::stri_count] [stri_match_all][stringi::stri_match_all]
}
