% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_funs.R
\name{read_nmlist}
\alias{read_nmlist}
\title{Import Utility for NONMEM xml files}
\usage{
read_nmlist(run, project, rundir = file.path(project, run))
}
\arguments{
\item{run}{character, run number(s) to parse}

\item{project}{character, root path of the NONMEM run(s)}

\item{rundir}{character, path containing the project, Default: file.path(project, run)}
}
\value{
xml_document class
}
\description{
Converts xml files to xml_document class
}
\examples{
read_nmlist(run = c('101'), project = system.file('streams',package = 'streamline'))
}
\seealso{
[read_xml][xml2::read_xml]
}
