% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_reactive.R
\name{validate_comments}
\alias{validate_comments}
\title{Reactive Validation}
\usage{
validate_comments(project = system.file("extdata", package = "tidynm"))
}
\arguments{
\item{project}{character, path to project containing runs, 
Default: system.file('extdata',package = 'tidynm')}
}
\description{
NONMEM Control Stream Comment Validation
}
\examples{
\dontrun{
if(interactive()){
 validate_comments()
 }
}
}
